[pdex] Add data structure for handling pending reward liqudity

+++ blockchain/pdex/state_v2_utils.go
@@ -3,7 +3,6 @@ package pdex
 import (
 	"encoding/json"
 	"math/big"
-	"reflect"
 
 	"github.com/incognitochain/incognito-chain/blockchain/pdex/v2utils"
 	"github.com/incognitochain/incognito-chain/common"
@@ -11,18 +10,32 @@ import (
 )
 
 type Share struct {
-	amount             uint64
-	tradingFees        map[common.Hash]uint64
-	lastLPFeesPerShare map[common.Hash]*big.Int
+	amount                uint64
+	lmLockedAmount        uint64
+	tradingFees           map[common.Hash]uint64
+	lastLPFeesPerShare    map[common.Hash]*big.Int
+	lastLmRewardsPerShare map[common.Hash]*big.Int
 }
 
 func (share *Share) Amount() uint64 {
 	return share.amount
 }
 
+func (share *Share) LmLockedShareAmount() uint64 {
+	return share.lmLockedAmount
+}
+
 func (share *Share) LastLPFeesPerShare() map[common.Hash]*big.Int {
 	res := make(map[common.Hash]*big.Int)
 	for k, v := range share.lastLPFeesPerShare {
+		res[k] = big.NewInt(0).Set(v)
+	}
+	return res
+}
+
+func (share *Share) LastLmRewardsPerShare() map[common.Hash]*big.Int {
+	res := make(map[common.Hash]*big.Int)
+	for k, v := range share.lastLmRewardsPerShare {
 		res[k] = big.NewInt(0).Set(v)
 	}
 	return res
@@ -38,27 +51,31 @@ func (share *Share) TradingFees() map[common.Hash]uint64 {
 
 func NewShare() *Share {
 	return &Share{
-		amount:             0,
-		tradingFees:        map[common.Hash]uint64{},
-		lastLPFeesPerShare: map[common.Hash]*big.Int{},
+		amount:                0,
+		tradingFees:           map[common.Hash]uint64{},
+		lastLPFeesPerShare:    map[common.Hash]*big.Int{},
+		lastLmRewardsPerShare: map[common.Hash]*big.Int{},
 	}
 }
 
 func NewShareWithValue(
-	amount uint64,
+	amount, lmLockedAmount uint64,
 	tradingFees map[common.Hash]uint64,
-	lastLPFeesPerShare map[common.Hash]*big.Int,
+	lastLPFeesPerShare, lastLmRewardsPerShare map[common.Hash]*big.Int,
 ) *Share {
 	return &Share{
-		amount:             amount,
-		tradingFees:        tradingFees,
-		lastLPFeesPerShare: lastLPFeesPerShare,
+		amount:                amount,
+		lmLockedAmount:        lmLockedAmount,
+		tradingFees:           tradingFees,
+		lastLPFeesPerShare:    lastLPFeesPerShare,
+		lastLmRewardsPerShare: lastLmRewardsPerShare,
 	}
 }
 
 func (share *Share) Clone() *Share {
 	res := NewShare()
 	res.amount = share.amount
+	res.lmLockedAmount = share.lmLockedAmount
 	res.tradingFees = map[common.Hash]uint64{}
 	for k, v := range share.tradingFees {
 		res.tradingFees[k] = v
@@ -67,18 +84,26 @@ func (share *Share) Clone() *Share {
 	for k, v := range share.lastLPFeesPerShare {
 		res.lastLPFeesPerShare[k] = new(big.Int).Set(v)
 	}
+	res.lastLmRewardsPerShare = map[common.Hash]*big.Int{}
+	for k, v := range share.lastLmRewardsPerShare {
+		res.lastLmRewardsPerShare[k] = new(big.Int).Set(v)
+	}
 	return res
 }
 
 func (share *Share) MarshalJSON() ([]byte, error) {
 	data, err := json.Marshal(struct {
-		Amount             uint64                   `json:"Amount"`
-		TradingFees        map[common.Hash]uint64   `json:"TradingFees"`
-		LastLPFeesPerShare map[common.Hash]*big.Int `json:"LastLPFeesPerShare"`
+		Amount                uint64                   `json:"Amount"`
+		TradingFees           map[common.Hash]uint64   `json:"TradingFees"`
+		LastLPFeesPerShare    map[common.Hash]*big.Int `json:"LastLPFeesPerShare"`
+		LmLockedAmount        uint64                   `json:"LmLockedAmount,omitempty"`
+		LastLmRewardsPerShare map[common.Hash]*big.Int `json:"LastLmRewardsPerShare,omitempty"`
 	}{
-		Amount:             share.amount,
-		TradingFees:        share.tradingFees,
-		LastLPFeesPerShare: share.lastLPFeesPerShare,
+		Amount:                share.amount,
+		TradingFees:           share.tradingFees,
+		LastLPFeesPerShare:    share.lastLPFeesPerShare,
+		LmLockedAmount:        share.lmLockedAmount,
+		LastLmRewardsPerShare: share.lastLmRewardsPerShare,
 	})
 	if err != nil {
 		return []byte{}, err
@@ -88,17 +113,21 @@ func (share *Share) MarshalJSON() ([]byte, error) {
 
 func (share *Share) UnmarshalJSON(data []byte) error {
 	temp := struct {
-		Amount             uint64                   `json:"Amount"`
-		TradingFees        map[common.Hash]uint64   `json:"TradingFees"`
-		LastLPFeesPerShare map[common.Hash]*big.Int `json:"LastLPFeesPerShare"`
+		Amount                uint64                   `json:"Amount"`
+		TradingFees           map[common.Hash]uint64   `json:"TradingFees"`
+		LastLPFeesPerShare    map[common.Hash]*big.Int `json:"LastLPFeesPerShare"`
+		LmLockedAmount        uint64                   `json:"LmLockedAmount,omitempty"`
+		LastLmRewardsPerShare map[common.Hash]*big.Int `json:"LastLmRewardsPerShare,omitempty"`
 	}{}
 	err := json.Unmarshal(data, &temp)
 	if err != nil {
 		return err
 	}
 	share.amount = temp.Amount
+	share.lmLockedAmount = temp.LmLockedAmount
 	share.tradingFees = temp.TradingFees
 	share.lastLPFeesPerShare = temp.LastLPFeesPerShare
+	share.lastLmRewardsPerShare = temp.LastLmRewardsPerShare
 	return nil
 }
 
@@ -118,12 +147,16 @@ func (share *Share) getDiff(
 		for tokenID := range share.lastLPFeesPerShare {
 			newShareChange.LastLPFeesPerShare[tokenID.String()] = true
 		}
+		for tokenID := range share.lastLmRewardsPerShare {
+			newShareChange.LastLmRewardsPerShare[tokenID.String()] = true
+		}
 	} else {
-		if !reflect.DeepEqual(share.amount, compareShare.amount) {
+		if share.amount != compareShare.amount || share.lmLockedAmount != compareShare.lmLockedAmount {
 			newShareChange.IsChanged = true
 		}
 		newShareChange.TradingFees = v2utils.GetChangedElementsFromMapUint64(share.tradingFees, compareShare.tradingFees)
 		newShareChange.LastLPFeesPerShare = v2utils.GetChangedElementsFromMapBigInt(share.lastLPFeesPerShare, compareShare.lastLPFeesPerShare)
+		newShareChange.LastLmRewardsPerShare = v2utils.GetChangedElementsFromMapBigInt(share.lastLmRewardsPerShare, compareShare.lastLmRewardsPerShare)
 	}
 
 	return newShareChange
@@ -253,9 +286,10 @@ func (share *Share) updateToDB(
 	if shareChange.IsChanged {
 		nftID, err := common.Hash{}.NewHashFromStr(nftID)
 		err = statedb.StorePdexv3Share(
-			env.StateDB(), poolPairID,
-			*nftID,
-			share.amount, share.tradingFees, share.lastLPFeesPerShare,
+			env.StateDB(), poolPairID, *nftID,
+			statedb.NewPdexv3ShareStateWithValue(
+				*nftID, share.amount, share.lmLockedAmount,
+			),
 		)
 		if err != nil {
 			return err
@@ -300,6 +334,24 @@ func (share *Share) updateToDB(
 					return err
 				}
 			} else {
+				err := statedb.DeletePdexv3ShareLastLpFeePerShare(
+					env.StateDB(), poolPairID, nftID, tokenID,
+				)
+				if err != nil {
+					return err
+				}
+			}
+			if lastLmRewardsPerShare, found := share.lastLmRewardsPerShare[*tokenHash]; found {
+				//TODO: @tin store to db
+				err := statedb.StorePdexv3ShareLastLpFeePerShare(
+					env.StateDB(), poolPairID, nftID,
+					statedb.NewPdexv3ShareLastLpFeePerShareStateWithValue(*tokenHash, lastLmRewardsPerShare),
+				)
+				if err != nil {
+					return err
+				}
+			} else {
+				//TODO: @tin delete from db
 				err := statedb.DeletePdexv3ShareLastLpFeePerShare(
 					env.StateDB(), poolPairID, nftID, tokenID,
 				)
